requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
magic-filter==1.0.12
multidict==6.6.3
packaging==25.0
phonenumbers==9.0.8
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
whitenoise==6.9.0
yarl==1.20.1
django-redis==6.0.0
